test06 complete, test07 almost complete
Only problem with test07 now is that for some reason I can't get Ctrl-C when a backgrounded process is put first, then the foreground process. May be because Ctrl-C is in built_in_commands, which are only meant to register when there isn't a foreground job. Need to find a new place to put Ctrl-C, Ctrl-Z and such