test07 working partially, working on test08
Test07 works not with the .txt file, but if I enter everything into the shell by hand. Test08 is having trouble that it doesnt go to the sigtstp_handler() when ctrl-z is pressed, it just sits in the process and does nothing. My guess is ctrl-z is stopping my shell.