Test07-Test13 working
From test07 to test14, everything is working in my shell, still not when the trace functions are used, which is odd. But happy I got them working. Will probably work on getting 14, 15, and 16 working before I find out why my shell doesn't like the trace files as I think it's a small issue.