Test07-Test14 now work with Trace functions
Waitfg() was the problem, or the lack thereof. Didn't write one until just now and now all those tests work. However, there is still a small issue with foreground processes processing information in a slightly different order (test14) and keeping the "/bin/echo tsh> jobs" (test11) but it only keeps that in test11, not in any other test.