Final Submission
Decided the extra 1-2 hours on test16 was not worth it and decided to take the 95. Everything else works perfectly. This update is added the error handling to all system calls from csapp.c and adding all the functions I needed into tsh.c. For some reason Execve was returning an error and execve did not so I decided to just use execve instead.