Completed assingment4
After many hours of debugging the issue I had was fixed. For some reason, when assigning a pointer in memory to a variable and then assigning the value of that variable to something, the code would compile funny and would result in seg faults. After assigning the value straight to the memory address, the issue was solved and my code compiled fine. End result was 90/100 for efficiency reading which is above the required 70/100 to get full credit.