Commands 1-5 work
Surprisingly my program worked for the fifth command before programming directly for it. A recent addition to assignment 5 is that I've found a way to make my program work in eval; instead of writing all my conditionals in builtin_cmd, I write in after the fork in eval. That makes it so the code is forked and everything else is taken care of. It's cut my code down by about 80%, most of which was boiler plate. Now all I have to do is make single pipe work, and if I feel like it, multiple pipes (multiple pipes is extra credit).